refactor: :art: useToggleSubscribeConcert
