--- apps/billets-app/src/features/subscribe/hooks/index.ts
@@ -1,1 +1,0 @@
-export * from './use-toggle-subscribe-concert'
