+++ apps/billets-app/src/features/subscribe/index.ts
@@ -1,2 +1,1 @@
-export * from './hooks'
 export * from './ui'
