+++ apps/billets-app/src/ui/venue-detail-concert-list/venue-detail-concert-list.tsx
@@ -1,4 +1,4 @@
-import { useToggleSubscribeConcert } from '@/features/subscribe'
+import { useToggleSubscribeConcert } from '@/features/subscribe/hooks/useToggleSubscribeConcert'
 import { apiClient } from '@/lib/api/openapi-client'
 import { useVenueDetailScreenNavigation } from '@/screens/venue-detail-screen/venue-detail-screen.hooks'
 import { useQuery, useSuspenseQuery } from '@tanstack/react-query'
